Make building independent of the Bundle
It's not necessary to use a `Bundle` to initialize a `Cluster` instance,
as it doesn't hook in to the server (e.g. Jersey) like Hibernate does.

Instead, it's better to have the `CassandraFactory` itself take the
`Environment` and be responsible for managing, adding healthchecks and
registering metrics.

I've preserved the `Bundle` API, to ensure existing code works.

Note: tests updated and pass. `README.md` updated to reflect new API.