Add build() overload that doesn't use Environment
Sometimes, you will want to configure and build a `Cluster` instance
using `CassandraFactory`, but you won't have an `Environment` available.

The common-case here is when writing a `ConfiguredCommand` that uses
Cassandra (e.g. for a utility to execute your applications' DDL
statements).

While we have no `Environment`, we can still register health-checks and
metrics, but we cannot manage the Cluster lifecycle, as there's no
Environment to tie its lifecycle to.

This is broadly following the idiom used by the Liquibase Commands in
`dropwizard-migrations` and the `DataSourceFactory` it uses to configure
connections.