Changed CassandraHealthCheck to use a validationQuery instead of Host.isUp
It seems the Cassandra driver is far too slow to determine host availability for the purposes of a health check, leading to false positives. This has been replaced with a validation query (with a default value) that is executed against the cluster. This is very reliable as the query only succeeds if it can be executed against a node in the cluster.