Added support for load balancing policies
The concept behind this was initially submitted in PR 12 (https://github.com/stuartgunter/dropwizard-cassandra/pull/12). This implementation is based on the standard pattern of typed configuration rather than reflection. Many thanks to adejanovski for the original suggestion and for taking the time to complete the PR that sparked this change.