Do not produce unnecessary log entry when instantiating an ObjectMapper.
Every time an ObjectMapper is instantiated while Dropwizard Cassandra
is on the classpath, Jackson produces the following log entry:

 INFO io.dropwizard.jackson.DiscoverableSubtypeResolver: Unable to load

This happens because Dropwizard’s DiscoverableSubtypeResolver assumes
each line in every file in META-INF/services is a valid fully qualified
class name and attempts to load it. When it’s not able to load a class
with an empty string as its name, it produces the log entry.